Fix --date option and add --config option
The --date option now only accepts dates formatted as YYYY-MM-DD and
converts them to the right format according to the config file. This
change allows us to read the path to the config file from the command
line since there's no longer a dependency on that file for the date
format.